Refactor: Extract file attachments to a new component
This change moves the file attachment display logic into a dedicated component, improving code organization and reusability.
